+++ fdb-record-layer-core/src/main/java/com/apple/foundationdb/record/metadata/expressions/FunctionKeyExpression.java
@@ -371,4 +371,18 @@ public abstract class FunctionKeyExpression extends BaseKeyExpression implements
             return functions;
         }
     }
+
+    @Override
+    public boolean equivalentForSort(@Nonnull KeyExpression other) {
+        if (getClass() != other.getClass()) {
+            return false;
+        }
+
+        FunctionKeyExpression that = (FunctionKeyExpression) other;
+        if (!this.getName().equals(that.getName())) {
+            return false;
+        }
+
+        return this.getArguments().equivalentForSort(that.getArguments());
+    }
 }