+++ fdb-record-layer-core/src/test/java/com/apple/foundationdb/record/provider/foundationdb/query/FDBSortQueryIndexSelectionTest.java
@@ -30,7 +30,9 @@ import com.apple.foundationdb.record.TestHelpers;
 import com.apple.foundationdb.record.TestRecords1Proto;
 import com.apple.foundationdb.record.TestRecordsWithHeaderProto;
 import com.apple.foundationdb.record.metadata.Index;
+import com.apple.foundationdb.record.metadata.IndexOptions;
 import com.apple.foundationdb.record.metadata.IndexTypes;
+import com.apple.foundationdb.record.metadata.Key;
 import com.apple.foundationdb.record.metadata.expressions.KeyExpression;
 import com.apple.foundationdb.record.metadata.expressions.KeyExpression.FanType;
 import com.apple.foundationdb.record.provider.foundationdb.FDBQueriedRecord;
@@ -123,9 +125,13 @@ public class FDBSortQueryIndexSelectionTest extends FDBRecordStoreQueryTestBase 
      * Verify that simple sorts are implemented using index scans.
      */
     @DualPlannerTest
-    public void sort() throws Exception {
+    public void sortOnly() throws Exception {
+        sortOnlyUnique(NO_HOOK);
+    }
+
+    private void sortOnlyUnique(RecordMetaDataHook hook) throws Exception {
         try (FDBRecordContext context = openContext()) {
-            openSimpleRecordStore(context);
+            openSimpleRecordStore(context, hook);
 
             for (int i = 0; i < 100; i++) {
                 TestRecords1Proto.MySimpleRecord.Builder recBuilder = TestRecords1Proto.MySimpleRecord.newBuilder();
@@ -145,7 +151,7 @@ public class FDBSortQueryIndexSelectionTest extends FDBRecordStoreQueryTestBase 
         assertEquals(-1130465929, plan.planHash());
 
         try (FDBRecordContext context = openContext()) {
-            openSimpleRecordStore(context);
+            openSimpleRecordStore(context, hook);
             int i = 0;
             try (RecordCursorIterator<FDBQueriedRecord<Message>> cursor = recordStore.executeQuery(plan).asIterator()) {
                 while (cursor.hasNext()) {
@@ -782,4 +788,64 @@ public class FDBSortQueryIndexSelectionTest extends FDBRecordStoreQueryTestBase 
         assertThat(plan, indexScan(allOf(indexName("schoolNameEmail"), bounds(hasTupleString("[[Human University],[Human University]]")))));
         assertEquals(387659205, plan.planHash());
     }
+
+    static final RecordMetaDataHook NULL_UNIQUE_HOOK = md -> {
+        md.removeIndex("MySimpleRecord$num_value_unique");
+        final Index index = new Index("MySimpleRecord$num_value_unique",
+                field("num_value_unique", FanType.None, Key.Evaluated.NullStandin.NULL_UNIQUE),
+                IndexTypes.VALUE, IndexOptions.UNIQUE_OPTIONS);
+        index.setSubspaceKey("MySimpleRecord$num_value_unique_2");  // Otherwise fails validation
+        md.addIndex("MySimpleRecord", index);
+    };
+
+    /**
+     * Verify that sort only works with different null interpretation on unique index.
+     */
+    @DualPlannerTest
+    public void sortOnlyUniqueNull() throws Exception {
+        sortOnlyUnique(NULL_UNIQUE_HOOK);
+    }
+    
+    /**
+     * Verify that sort with filter works with different null interpretation on unique index.
+     */
+    @DualPlannerTest
+    public void sortUniqueNull() throws Exception {
+        try (FDBRecordContext context = openContext()) {
+            openSimpleRecordStore(context, NULL_UNIQUE_HOOK);
+
+            for (int i = 0; i < 100; i++) {
+                TestRecords1Proto.MySimpleRecord.Builder recBuilder = TestRecords1Proto.MySimpleRecord.newBuilder();
+                recBuilder.setRecNo((1096 * i + 722) % 1289);
+                recBuilder.setNumValueUnique(i);
+                recordStore.saveRecord(recBuilder.build());
+            }
+            commit(context);
+        }
+
+        RecordQuery query = RecordQuery.newBuilder()
+                .setRecordType("MySimpleRecord")
+                .setFilter(Query.field("num_value_unique").greaterThanOrEquals(20))
+                .setSort(field("num_value_unique"))
+                .build();
+        RecordQueryPlan plan = planner.plan(query);
+        assertThat(plan, indexScan(allOf(indexName("MySimpleRecord$num_value_unique"), bounds(hasTupleString("[[20],>")))));
+        assertEquals(-535398101, plan.planHash());
+
+        try (FDBRecordContext context = openContext()) {
+            openSimpleRecordStore(context, NULL_UNIQUE_HOOK);
+            int i = 20;
+            try (RecordCursorIterator<FDBQueriedRecord<Message>> cursor = recordStore.executeQuery(plan).asIterator()) {
+                while (cursor.hasNext()) {
+                    FDBQueriedRecord<Message> rec = cursor.next();
+                    TestRecords1Proto.MySimpleRecord.Builder myrec = TestRecords1Proto.MySimpleRecord.newBuilder();
+                    myrec.mergeFrom(rec.getRecord());
+                    assertEquals(i++, myrec.getNumValueUnique());
+                }
+            }
+            assertEquals(100, i);
+            assertDiscardedNone(context);
+        }
+    }
+
 }